# clients/apple/Packages/TorrentCoreKit/Sources/TorrentCoreAPI/openapi.json
{
  "components": {
    "schemas": {
...
      "AddMagnetRequest": {
        "additionalProperties": false,
        "properties": {
          "categoryKey": {
            "nullable": true,
            "type": "string"
          },
          "magnetUri": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "magnetUri"
        ],
        "type": "object"
      },
...
      "RemoveTorrentRequest": {
        "additionalProperties": false,
        "properties": {
          "deleteData": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ReportCompletionCallbackResultRequest": {
        "additionalProperties": false,
        "properties": {
          "allowResubmit": {
            "type": "boolean"
          },
          "attemptCount": {
            "format": "int32",
            "type": "integer"
          },
          "callbackFinished": {
            "type": "boolean"
          },
          "callbackLocalTimestamp": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "callbackMachine": {
            "nullable": true,
            "type": "string"
          },
          "callbackSource": {
            "nullable": true,
            "type": "string"
          },
          "completionTimestamp": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "contractVersion": {
            "nullable": true,
            "type": "string"
          },
          "correlationId": {
            "nullable": true,
            "type": "string"
          },
          "detailMessage": {
            "nullable": true,
            "type": "string"
          },
          "displayMessage": {
            "nullable": true,
            "type": "string"
          },
          "finalState": {
            "nullable": true,
            "type": "string"
          },
          "mediaConsideredDone": {
            "type": "boolean"
          },
          "needsManualIntervention": {
            "type": "boolean"
          },
          "rawResponseJson": {
            "nullable": true,
            "type": "string"
          },
          "reasonCode": {
            "nullable": true,
            "type": "string"
          },
          "recommendedAction": {
            "nullable": true,
            "type": "string"
          },
          "resubmitAdvice": {
            "nullable": true,
            "type": "string"
          },
          "sourceState": {
            "nullable": true,
            "type": "string"
          },
          "torrentHash": {
            "nullable": true,
            "type": "string"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "torrentId"
        ],
        "type": "object"
      },
...
      "ServiceProblemDetailsDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "nullable": true,
            "type": "string"
          },
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "nullable": true,
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "target": {
            "nullable": true,
            "type": "string"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "traceId": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TorrentActionResultDto": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "nullable": true,
            "type": "string"
          },
          "dataDeleted": {
            "type": "boolean"
          },
          "processedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TorrentState"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "action",
          "processedAtUtc",
          "state",
          "torrentId"
        ],
        "type": "object"
      },
...
      "TorrentDetailDto": {
        "additionalProperties": false,
        "properties": {
          "addedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "canPause": {
            "type": "boolean"
          },
          "canRefreshMetadata": {
            "type": "boolean"
          },
          "canRemove": {
            "type": "boolean"
          },
          "canResume": {
            "type": "boolean"
          },
          "canRetryCompletionCallback": {
            "type": "boolean"
          },
          "categoryKey": {
            "nullable": true,
            "type": "string"
          },
          "completedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackFeedback": {
            "$ref": "#/components/schemas/CompletionCallbackFeedbackDto"
          },
          "completionCallbackFinalPayloadPath": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackInvokedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackLastError": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackPendingReason": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackPendingSinceUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackState": {
            "nullable": true,
            "type": "string"
          },
          "connectedPeerCount": {
            "format": "int32",
            "type": "integer"
          },
          "downloadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "downloadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "errorMessage": {
            "nullable": true,
            "type": "string"
          },
          "infoHash": {
            "nullable": true,
            "type": "string"
          },
          "lastActivityAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "magnetUri": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "progressPercent": {
            "format": "double",
            "type": "number"
          },
          "queuePosition": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "savePath": {
            "nullable": true,
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TorrentState"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          },
          "totalBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "trackerCount": {
            "format": "int32",
            "type": "integer"
          },
          "uploadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "waitReason": {
            "$ref": "#/components/schemas/TorrentWaitReason"
          }
        },
        "required": [
          "addedAtUtc",
          "canPause",
          "canRefreshMetadata",
          "canRemove",
          "canResume",
          "canRetryCompletionCallback",
          "connectedPeerCount",
          "downloadedBytes",
          "downloadRateBytesPerSecond",
          "magnetUri",
          "name",
          "progressPercent",
          "savePath",
          "state",
          "torrentId",
          "trackerCount",
          "uploadRateBytesPerSecond"
        ],
        "type": "object"
      },
...
      "TorrentState": {
        "enum": [
          "ResolvingMetadata",
          "Queued",
          "Downloading",
          "Seeding",
          "WaitingForFileCompletion",
          "Paused",
          "Completed",
          "Error",
          "Removed"
        ],
        "type": "string"
      },
      "TorrentSummaryDto": {
        "additionalProperties": false,
        "properties": {
          "addedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "canPause": {
            "type": "boolean"
          },
          "canRefreshMetadata": {
            "type": "boolean"
          },
          "canRemove": {
            "type": "boolean"
          },
          "canResume": {
            "type": "boolean"
          },
          "canRetryCompletionCallback": {
            "type": "boolean"
          },
          "categoryKey": {
            "nullable": true,
            "type": "string"
          },
          "completedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackInvokedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackLastError": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackPendingSinceUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completionCallbackState": {
            "nullable": true,
            "type": "string"
          },
          "connectedPeerCount": {
            "format": "int32",
            "type": "integer"
          },
          "downloadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "downloadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "errorMessage": {
            "nullable": true,
            "type": "string"
          },
          "lastActivityAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "progressPercent": {
            "format": "double",
            "type": "number"
          },
          "queuePosition": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/TorrentState"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          },
          "totalBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "trackerCount": {
            "format": "int32",
            "type": "integer"
          },
          "uploadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "waitReason": {
            "$ref": "#/components/schemas/TorrentWaitReason"
          }
        },
        "required": [
          "addedAtUtc",
          "canPause",
          "canRefreshMetadata",
          "canRemove",
          "canResume",
          "canRetryCompletionCallback",
          "connectedPeerCount",
          "downloadedBytes",
          "downloadRateBytesPerSecond",
          "name",
          "progressPercent",
          "state",
          "torrentId",
          "trackerCount",
          "uploadRateBytesPerSecond"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Management API for the TorrentCore service host.",
    "title": "TorrentCore Service API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/torrents": {
      "get": {
        "operationId": "Torrents_GetAll",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TorrentSummaryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Torrents"
        ]
      },
      "post": {
        "operationId": "Torrents_Add",
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/AddMagnetRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMagnetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMagnetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentDetailDto"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}": {
      "get": {
        "operationId": "Torrents_GetById",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentDetailDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/completion-callback/result": {
      "post": {
        "operationId": "Torrents_ReportCompletionCallbackResult",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ReportCompletionCallbackResultRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCompletionCallbackResultRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCompletionCallbackResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/completion-callback/retry": {
      "post": {
        "operationId": "Torrents_RetryCompletionCallback",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/metadata/refresh": {
      "post": {
        "operationId": "Torrents_RefreshMetadata",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/metadata/reset": {
      "post": {
        "operationId": "Torrents_ResetMetadataSession",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/pause": {
      "post": {
        "operationId": "Torrents_Pause",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/peers": {
      "get": {
        "operationId": "Torrents_GetPeers",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TorrentPeerDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/remove": {
      "post": {
        "operationId": "Torrents_Remove",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTorrentRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTorrentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTorrentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/resume": {
      "post": {
        "operationId": "Torrents_Resume",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentActionResultDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    },
    "/api/torrents/{torrentId}/trackers": {
      "get": {
        "operationId": "Torrents_GetTrackers",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TorrentTrackerDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "Torrents"
        ]
      }
    }
  },
...
}