# src/TorrentCore.ServiceHost/Controllers/TorrentsController.cs
#region

using Microsoft.AspNetCore.Mvc;
using TorrentCore.Contracts;
using TorrentCore.Contracts.Torrents;
using TorrentCore.Service.Application;

#endregion

namespace TorrentCore.Service.Controllers;

[ApiController]
[Route("api/torrents")]
[Produces("application/json")]
public sealed class TorrentsController(ITorrentApplicationService torrentApplicationService) : ControllerBase
{
    [HttpGet]
    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IReadOnlyList<TorrentSummaryDto>))]
    public async Task<ActionResult<IReadOnlyList<TorrentSummaryDto>>> GetAll(CancellationToken cancellationToken)
    {
        var torrents = await torrentApplicationService.GetTorrentsAsync(cancellationToken);
        return Ok(torrents);
    }

    [HttpGet("{torrentId:guid}")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentDetailDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentDetailDto>> GetById(Guid torrentId, CancellationToken cancellationToken)
    {
        var torrent = await torrentApplicationService.GetTorrentAsync(torrentId, cancellationToken);
        return Ok(torrent);
    }

    [HttpGet("{torrentId:guid}/peers")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(IReadOnlyList<TorrentPeerDto>))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<IReadOnlyList<TorrentPeerDto>>> GetPeers(Guid torrentId,
        CancellationToken                                                      cancellationToken)
    {
        var peers = await torrentApplicationService.GetTorrentPeersAsync(torrentId, cancellationToken);
        return Ok(peers);
    }

    [HttpGet("{torrentId:guid}/trackers")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(IReadOnlyList<TorrentTrackerDto>))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<IReadOnlyList<TorrentTrackerDto>>> GetTrackers(Guid torrentId,
        CancellationToken                                                         cancellationToken)
    {
        var trackers = await torrentApplicationService.GetTorrentTrackersAsync(torrentId, cancellationToken);
        return Ok(trackers);
    }

    [HttpPost]
    [ProducesResponseType(StatusCodes.Status201Created,    Type = typeof(TorrentDetailDto))]
    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict,   Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status503ServiceUnavailable, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentDetailDto>> Add([FromBody] AddMagnetRequest request,
        CancellationToken                                                             cancellationToken)
    {
        var torrent = await torrentApplicationService.AddMagnetAsync(request, cancellationToken);
        return CreatedAtAction(nameof(GetById), new {torrentId = torrent.TorrentId}, torrent);
    }

    [HttpPost("{torrentId:guid}/pause")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> Pause(Guid torrentId, CancellationToken cancellationToken)
    {
        var result = await torrentApplicationService.PauseAsync(torrentId, cancellationToken);
        return Ok(result);
    }

    [HttpPost("{torrentId:guid}/resume")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> Resume(Guid torrentId, CancellationToken cancellationToken)
    {
        var result = await torrentApplicationService.ResumeAsync(torrentId, cancellationToken);
        return Ok(result);
    }

    [HttpPost("{torrentId:guid}/metadata/refresh")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> RefreshMetadata(Guid torrentId,
        CancellationToken                                                        cancellationToken)
    {
        var result = await torrentApplicationService.RefreshMetadataAsync(torrentId, cancellationToken);
        return Ok(result);
    }

    [HttpPost("{torrentId:guid}/metadata/reset")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> ResetMetadataSession(Guid torrentId,
        CancellationToken                                                             cancellationToken)
    {
        var result = await torrentApplicationService.ResetMetadataSessionAsync(torrentId, cancellationToken);
        return Ok(result);
    }

    [HttpPost("{torrentId:guid}/completion-callback/retry")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> RetryCompletionCallback(Guid torrentId,
        CancellationToken                                                                cancellationToken)
    {
        var result = await torrentApplicationService.RetryCompletionCallbackAsync(torrentId, cancellationToken);
        return Ok(result);
    }

    [HttpPost("{torrentId:guid}/completion-callback/result")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(ServiceProblemDetailsDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound,   Type = typeof(ServiceProblemDetailsDto))]
    public async Task<IActionResult> ReportCompletionCallbackResult(Guid torrentId,
        [FromBody] ReportCompletionCallbackResultRequest request, CancellationToken cancellationToken)
    {
        await torrentApplicationService.ReportCompletionCallbackResultAsync(torrentId, request, cancellationToken);
        return Ok();
    }

    [HttpPost("{torrentId:guid}/remove")]
    [ProducesResponseType(StatusCodes.Status200OK,       Type = typeof(TorrentActionResultDto))]
    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ServiceProblemDetailsDto))]
    public async Task<ActionResult<TorrentActionResultDto>> Remove(Guid torrentId,
        [FromBody] RemoveTorrentRequest?                                request, CancellationToken cancellationToken)
    {
        var result = await torrentApplicationService.RemoveAsync(
            torrentId, request ?? new RemoveTorrentRequest(), cancellationToken
        );
        return Ok(result);
    }
}